Permit running backwards-in-time
Specifying a negative etaf parameter now permits running the ODE
backward in time.  This is handy to get a couple of Blasius solution
points prior to eta = 0 so that interpolation routines that set
natural end conditions can be fooled with data prior to the region
of interest we genuinely want.